extract utils
